Create TrainLine component as a base for StraightTrainLine
