+++ web-app/components/StraightTrainLine.tsx
@@ -1,42 +1,14 @@
-import { Circle, Polyline } from '@react-google-maps/api'
 import { Point } from '../types/points'
+import TrainLine from './TrainLine'
 
-interface TrainLineProps {
+interface StraightTrainLineProps {
   points: Point[]
   color: string
 }
 
-const StraightTrainLine = ({ points, color }: TrainLineProps) => {
-  return (
-    <>
-      {points && points.length > 0 && (
-        <>
-          {points.map((point, index) => (
-            <Circle
-              key={index}
-              center={point}
-              radius={50}
-              options={{
-                fillColor: 'white',
-                fillOpacity: 1,
-                strokeColor: 'black',
-                strokeOpacity: 1,
-                strokeWeight: 3,
-              }}
-            />
-          ))}
-          <Polyline
-            path={points}
-            options={{
-              strokeColor: color,
-              strokeOpacity: 1,
-              strokeWeight: 5,
-            }}
-          />
-        </>
-      )}
-    </>
-  )
+const StraightTrainLine = ({ points, color }: StraightTrainLineProps) => {
+  const paths = [points]
+  return <TrainLine points={points} color={color} paths={paths} />
 }
 
 export default StraightTrainLine
